requirements.txt
flask
psycopg2
Flask-Cors

Flask-SQLAlchemy
